pkg/csilvm: don't run pvck and vgck on Probe
It appears that these commands scan the PVs and rebuilds / modifies the LVM2
metadata stored on the disk. The pvck and vgck commands expect user input.

We can run with --yes which answers yes to all prompts (which the man page
warns against with 'extreme caution').

We can run with --test which avoids updating metadata but still responds with
success. The man page warns that this may yield to confusing errors when
multi-stage operations expect that metadata was updated but it wasn't.

We can run with -qq which suppresses all output and answers 'no' to prompts.

None of these seem satisfying. We avoid running pvck and vgck until we
understand more about when to use it.